Check for thumbnails deletion
When a user deletes his/her account (and thus
its linked votes), it may happen that a survey
vote count falls under the threshold for
thumbnail generation, which means it can remain
with a wrong thumbnail. Now we check each survey
and if its vote count falls under that thershold,
we try to delete its thumbnails (no error raised
if they don't exist).