Handle session file reading errors
They are quite rare, but when they occur, the application
usually crashes. Now we set the session as expired when
it can't be read.
Also explicitly disabled the i18n-2 module debug messages.
There are a ton of them and they can easily clutter the
output, even in development mode.