Refactor Utils.checkUrlIsImage
Now we use the request module and also apply a
timeout in order to prevent a strange problem
that leaves some HTTPS connections hanging without
response. If the request is not answered in time,
image will be accepted but we'll log a warning
about it (maybe thr remote server is down
temporarily).